requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
pytesseract==0.3.10
PyMuPDF==1.23.8
Werkzeug==2.3.7 